delete extra files
