--- paddlevlp/models/evaclip/eva_clip/__init__.py
@@ -1,5 +1,0 @@
-from .constants import OPENAI_DATASET_MEAN, OPENAI_DATASET_STD
-from .loss import ClipLoss
-from .model import EVACLIP
-# from .coca_model import CoCa
-from .pretrained import list_pretrained, list_pretrained_models_by_tag, list_pretrained_tags_by_model, get_pretrained_url, download_pretrained_from_url, is_pretrained_cfg, get_pretrained_cfg, download_pretrained
