--- paddlevlp/models/evaclip/eva_clip/constants.py
@@ -1,2 +1,0 @@
-OPENAI_DATASET_MEAN = 0.48145466, 0.4578275, 0.40821073
-OPENAI_DATASET_STD = 0.26862954, 0.26130258, 0.27577711
